move files forked from eslint to `eslint/`
