+++ .eslintrc.cjs
@@ -30,12 +30,15 @@ module.exports = {
           // 親ディレクトリの同一ディレクトリにあるモジュールは許可
           '!../*.js',
 
-          // cli と util は別にディレクトリで境界づけたい動機もないので許可
+          // cli, util, eslint は別にディレクトリで境界づけたい動機もないので許可
           '!./cli/*',
           '!../cli/*',
           '!./util/*',
           '!../util/*',
           '!../../util',
+          '!./eslint/*',
+          '!../eslint/*',
+          '!../../eslint',
         ],
       },
     ],
