+++ src/eslint/report-translator.ts
