+++ src/fixer/fix/disable-per-line.ts
@@ -1,5 +1,6 @@
 import { Linter, Rule, SourceCode } from 'eslint';
 import { DescriptionPosition } from 'src/cli/prompt.js';
+import { mergeFixes } from '../../eslint/report-translator.js';
 import { groupBy, unique } from '../../util/array.js';
 import {
   DisableComment,
@@ -12,7 +13,6 @@ import {
 } from '../../util/eslint.js';
 import { notEmpty } from '../../util/type-check.js';
 import { FixContext } from '../index.js';
-import { mergeFixes } from '../report-translator.js';
 
 export type FixToDisablePerLineArgs = {
   description?: string | undefined;
